Add samurai archer enemy
The samurai archer is a stationary enemy. This is the first enemy that will intentionally target the player. It will constantly spin towards the players position, and has a simple bow drawing animation. Every time the animation finishes, a projectile will be spawned and fired towards that position. It spawns by gradually increasing it's own alpha until it's visible, which is when the collision mechanisms and animation sequences are turned on.