Added Lightning Trident weapon
Ok this weapon is just really cool. It works by sending a lightning shock to the nearest enemy, killing both the attacked enemy, the closest one via electrocution, and any other unfortunate enemies who happen to stand in your way. It also sounds really cool.

Does some magical functional programming to find the nearest enemy (could be expanded to have chain lightning?), then calculating the position of the start and end positions of the lightning, setting the rotation, and stretching the sprite to fit it. Also has a very short animation that consists of flipping on Y and X axis to simulate lightning "movement". Also has a Light2D component to light up the surroundings. Did I mention that this is really cool???