Add new level (warehouse), level obstructions, improved spawn logic
Warehouse level features two boxes in the middle, that obstruct movement from enemies, the player, as well as bullets. Hopefully bullets being obstructed will lead to the player being able to make more tactical plays by hiding behind boxes when there are many archers in one area directly opposite of the box, for example.

The box is an example of simple square shaped obstruction.

The new spawning logic is quite nice. It works by assigning a list of illegal gameobjects that contain colliders. Every time a spawner wants to Spawn() something, if the randomized location it picks is illegal, it should be moved over randomly to a hopefully legal area.

The way this works is pretty cool. First, I create a BoxCast gameobject which will be created in the location the enemy is supposed to spawn in. Then the code will loop over all of the illegal colliders, and if any of those are the same as the result of the BoxCast, a new location for the enemy will be chosen along a scaled down unit circle, this prevents a new random location still inside of a box. All spawn gameobjects will have a checkbox if they want to use this somewhat expensive operation, so it should only be done if there are actual obstructions on the map.