Added early stages of paintbrush enemy like art assets, not yet functional
Paintbrush will jump around leaving behind randomly colored spot of paint on the ground as they land. Walking on this paint will somehow debilitate the player. I'm thinking either damage (might be boring) or a crowd control effect of some sort, maybe the screen takes a slilght tint of that color.

Need to add movement script to it, as well as animation and collider that matches up with the sprite, and also the actual paint sprites that it leaves on the ground.