Add fully functioning Paintbrush enemy
Will spawn random colors on the ground where it lands, looks pretty cool. Fully animated and moves around similar to the turrets.

Need to do: move down color spawning slightly relative to the size of the paintbrush enemy
And make the color either go away with time or make it removable with shovel, not sure which is best yet