Paint spots now work like enemy remains. Paintbrush enemy is basically complete
Color spawns roughly on the middle of the brush hairs rather than the center of the brush enemy, looks more realistic.

Paint spots stay behind for a variable amount of time, using a refactored slime remains script now used for any enemy remains. Can also be removed with the shovel as intended.

Added some ideas on how the paint can be used other than just kill the player in todo.