Alter lvl 16, add victory theme sound
Lvl 16 was way too easy, I let the paintbrushes spawn more often and added in archers too.

The victory theme song will play by default when the gamemanager triggers the Win() function, it seems to work pretty well. Only problem is that the victory theme mp3 is much lower volume compared to the rest of the songs so I had to do some dirty code to increase it.