Add boss fight! Giant Void Slime. Also main menu button
Void Slime is large, scary and slippery. Will target the player and chase them at high speed. Takes several hits to kill. Can move out of bounds, because they are one with the void.

Main menu buttons: All in-game scenes now have a button that takes you back to the main menu instantly, instead of having to die or win to go there.